+++ tests/unit/GeneralizedQuantifier_test.go
@@ -62,8 +62,7 @@ func TestGeneralizedQuantifier(t *testing.T) {
 	systemFunctionBase := knowledge.NewSystemFunctionBase("system-function", log)
 	solver.AddFunctionBase(systemFunctionBase)
 	tokenizer := parse.NewTokenizer(parse.DefaultTokenizerExpression)
-	nameResolver := central.NewNameResolver(solver, meta, log, dialogContext)
-	parser := earley.NewParser(nameResolver, meta, log)
+	parser := earley.NewParser(log)
 
 	tests := []struct {
 		input string
@@ -84,7 +83,7 @@ func TestGeneralizedQuantifier(t *testing.T) {
 	for _, test := range tests {
 
 		words := tokenizer.Process(test.input)
-		trees := parser.Parse(grammarRules, words)
+		trees := parser.Parse(grammarRules, nil, words)
 		if len(trees) == 0 {
 			t.Errorf("Cannot parse: %s", test.input)
 			continue