+++ tests/unit/MorphologicalAnalyser_test.go
@@ -1,7 +1,6 @@
 package tests
 
 import (
-	"nli-go/lib/central"
 	"nli-go/lib/common"
 	"nli-go/lib/importer"
 	"nli-go/lib/morphology"
@@ -13,9 +12,8 @@ func TestSegmenter(t *testing.T) {
 
 	log := common.NewSystemLog()
 	parser := importer.NewInternalGrammarParser()
-	segmenter := morphology.NewSegmenter()
 
-	segmentationRules := parser.CreateSegmentationRulesAndCharacterClasses(`
+	segmentationRules := parser.CreateSegmentationRules(`
 		vowel: ['a', 'e', 'i', 'o', 'u', 'y']
 		consonant: ['b', 'c', 'd', 'f', 'g', 'h', 'j', 'k', 'l', 'm', 'n', 'p', 'q', 'r', 's', 't', 'v', 'w', 'x', 'z']
 		comp: '*{consonant1}{consonant1}er' -> adj: '*{consonant1}', suffix: 'er'
@@ -28,6 +26,7 @@ func TestSegmenter(t *testing.T) {
 		suffix: 'est'
 		suffix: 'er'
 	`)
+	segmenter := morphology.NewSegmenter(segmentationRules)
 
 	tests := []struct {
 		input string
@@ -42,43 +41,10 @@ func TestSegmenter(t *testing.T) {
 
 	for _, test := range tests {
 
-		result := strings.Join(segmenter.Segment(segmentationRules, test.input, test.cat), " ")
+		result := strings.Join(segmenter.Segment(test.input, test.cat), " ")
 
 		if result != test.want {
 			t.Errorf("call %v: got %v, want %v", test.input, result, test.want)
-		}
-	}
-
-	if len(log.GetErrors()) > 0 {
-		t.Errorf("errors: %v", log.String())
-	}
-}
-
-func TestMorphologicalAnalyser(t *testing.T) {
-
-	log := common.NewSystemLog()
-	parser := importer.NewInternalGrammarParser()
-	analyser := central.NewMorphologicalAnalyser()
-
-	tests := []struct {
-		input      string
-		cat     string
-		variables string
-		want string
-		ok bool
-	}{
-		{"biggest", "super", "E1", "height(E1, A1) order(A1, desc) first()", true},
-	}
-
-	for _, test := range tests {
-
-		want := parser.CreateRelationSet(test.want)
-		variables := strings.Split(test.variables, ",")
-
-		result, ok := analyser.Analyse(test.input, test.cat, variables)
-
-		if result.String() != want.String() || ok != test.ok {
-			t.Errorf("call %v: got %v, want %v", test.input, result.String(), want.String())
 		}
 	}